refactor(common): compile extent directory into single file
affects: @esri/hub-common
