+++ packages/common/src/extent.ts
@@ -1,7 +1,34 @@
-import { IHubRequestOptions } from "../types";
-import { getProp } from "../objects";
-import { IRequestOptions, request } from "@esri/arcgis-rest-request";
 import { IExtent } from "@esri/arcgis-rest-types";
+import { IHubRequestOptions, IBBox } from "./types";
+import { getProp } from "./objects";
+import { IRequestOptions, request } from "@esri/arcgis-rest-request";
+
+export function createExtent(
+  xmin: number,
+  ymin: number,
+  xmax: number,
+  ymax: number,
+  wkid: number = 4326
+): IExtent {
+  return {
+    xmin,
+    ymin,
+    xmax,
+    ymax,
+    // type: 'extent',
+    spatialReference: {
+      wkid
+    }
+  };
+}
+
+/**
+ * Turns an extent into a bbox
+ * @param envelope extent
+ */
+export function extentToBBox(envelope: IExtent): IBBox {
+  return [[envelope.xmin, envelope.ymin], [envelope.xmax, envelope.ymax]];
+}
 
 export const GLOBAL_EXTENT: IExtent = {
   xmin: -180,
@@ -67,4 +94,42 @@ export function getGeographicOrgExtent(
     .catch(ex => {
       return GLOBAL_EXTENT;
     });
+}
+
+/**
+ * Get the default org extent as a bbox for use on item.extent
+ * @param {IHubRequestOptions} hubRequestOptions
+ */
+export function getOrgExtentAsBBox(
+  hubRequestOptions: IHubRequestOptions
+): Promise<IBBox> {
+  return getGeographicOrgExtent(hubRequestOptions).then(extent =>
+    extentToBBox(extent)
+  );
+}
+
+export function isExtentCoordinateArray(extent: object) {
+  return (
+    Array.isArray(extent) &&
+    Array.isArray(extent[0]) &&
+    Array.isArray(extent[1])
+  );
+}
+
+function isExtentJSON(extent: any) {
+  return ["xmin", "ymin", "xmax", "ymax"].every(
+    key => typeof extent[key] === "number"
+  );
+}
+
+/**
+ * Check if the given extent is in a known format
+ * @param  {Object} extent extent in any format
+ * @return {Boolean}       indicator
+ */
+export function isValidExtent(extent: object) {
+  return (
+    !!extent &&
+    [isExtentCoordinateArray, isExtentJSON].some(test => test(extent))
+  );
 }
